Rename source code file
